fix: Use environment variables for custom base paths
To prevent the issue where navigation from inner routes would fail -
i.e. going to /mappings from /patronRequests would lead to the
non-existent /patronRequests/mappings - I have introduced a new
environment variable called VITE_PUBLIC_URL. If deploying to multiple
folders, we should now specify the base path here to avoid the
aforementioned issue.
This is a work in progress as it would be good to combine the two and
fix the issue while maintaining the dynamic approach from before.