feat: Add better error handling for filters and sorts [DCB-2100]
This addresses the issue whereby DCB Admin for Libraries users could
become stuck on a page if their filters or sorts returned an error.
Now we have a new hook that detects these errors, resets the local
state, warns the user, and allows them to try again.

I have also tightened up a few columns where sorting was presented as an
option and should not have been.